Change bindings to use libscrypt-kdf
This library originally used C source files for its scrypt
implementation. I don't know if they were home-rolled or copied from the
scrypt repository, but they were quite old and not portable.

This replaces those C files with a dependency on Tarsnap's own
`libscrypt-kdf` library, which defines a convenient exposed function for
calculating the scrypt hash of some bytes. This library should be newer
and more portable, so hopefully this makes the Haskell package usable on
multiple platforms (particularly M1 Macs, which are an obstacle at the
moment at Supercede.)

This also makes the repository into a Nix flake, to simplify using it at
Supercede and by Supercede developers.