Update project version and documentation
This adds some more up-to-date docs, in case we want to take over
Hackage maintenance for this particular library.

It also bumps the version, to help clarify when we're relying on our own
custom release on scrypt in other projects, and in case we want to take
over Hackage maintenance.