Use explicitly-generated cabal2nix file
For some reason, relying on `callCabal2nix` in the all-compilers overlay
wasn't working when I tried to depend on our modified `scrypt` package
for our own use - something about the resulting derivation was making
Nix unhappy. However, this manually-generated Nix derivation *does*
work, and is functionally identical, so let's use it. The only annoyance
will be remembering to re-generate the file when we change the `.cabal`
file.

This also avoids a use of IFD, which flakes don't tend to like.